Split test_rick
