+++ tests/testbot.py
@@ -64,19 +64,6 @@ class TestCase(TofbotTestCase):
         self.bot.send("oho")
         self.assertNoOutput("s/++/a")
 
-    @httprettified
-    def test_rick(self):
-        url = 'https://www.youtube.com/watch?v=dQw4w9WgXcQ'
-        title = 'rickroll'
-        response = '<html><head><title>%s</title></head></html>' % title
-        HTTPretty.register_uri(HTTPretty.GET, url,
-                               body=response,
-                               )
-        self.assertOutput("Keyboard cat v2: %s" % url,
-                          [title,
-                           "We're no strangers to love..."
-                           ])
-
     def test_like(self):
         self.assertOutputLength('!ggg', 0)
         self.bot.send('oh oh', origin='alfred')
